requirements.txt
flask
flask-cors-5.0.1
